go/runtime/txpool: move Transaction struct
